Add files via upload
This CSV file has "aspiration" column back, while 'clean_usedcars.csv" did not have aspiration column as we created a dummy variable for "aspiration" column as it has two values i)std, ii)turbo, for better analysis, but I realized that aspiration column is not that useful feature to predict the price of the car as it is not correlated with target variable price. The correlation value is -0.17 and 0.17 respectively, which is near by 0, indicating poor correlation between "aspiration" and "price".